moving the foreach processor to pipeline_processors
